Don't reexport everything in the root of libredox + Use less pointer sized integers (*size -> *32 in liborb{ital, tk})

+++ filesystem/apps/zfs/main.rs
@@ -1,5 +1,8 @@
 //To use this, please install zfs-fuse
-use std::{mem, str, File, Read, ToNum};
+use std::{mem, str};
+use std::to_num::ToNum;
+use std::fs::File;
+use std::io::Read;
 
 use self::arcache::ArCache;
 use self::dnode::{DNodePhys, ObjectSetPhys, ObjectType};
@@ -549,7 +552,7 @@ impl Zfs {
                                 let sector = arg.to_num();
                                 println!("Dump sector: {}", sector);
 
-                                let data = zfs.reader.zio.read(sector, 1);
+                                let data = zfs.reader.zio.read(sector as usize, 1);
                                 for i in 0..data.len() {
                                     if i % 32 == 0 {
                                         print!("\n{:X}:", i);
